Tho/fix order (#83)
* Fix total order and total discount

* Fix total order and total discount

+++ bff/admin_bff/schema/order_service/response.go
@@ -31,7 +31,7 @@ type OrderGoodsResponse struct {
 	Name      string
 	UnitPrice int
 	Price     int
-	Tax       int
+	Tax       float32
 	Quantity  int
 	Size      string
 	Color     string
